fix for findBySlug to work with Eager Loading
I have rebuild the findBySlug method using query scope so now it works with eager loading as well.

Previously this would not work! 
$content = Content::with('images','documents','videos','metatarsi')->findBySlug('page-1')->first();

However by using query scope to build the findBySlug method it works beautifully :)

P.S. why are you using ->get() and not terminating the query with ->first() in this method? Logically when you get a record by its slug its unique so its a single item! didn't change this since it would change existing behavior for anyone using your package but thought of mentioning this. tests would fail since i was using ->first() but quickly realized that you were using ->get() and reverted to that so that phpunit would not complain